backend/requirements.txt
# Web Framework
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
python-jose>=3.3.0
python-dotenv>=0.19.0

# Image Processing
pytesseract>=0.3.8
Pillow>=10.0.0
opencv-python>=4.8.0

# Document Processing
pdf2image>=1.16.0
python-docx>=0.8.11
openpyxl>=3.1.2  # Excel files
xlrd>=2.0.1     # Legacy Excel (.xls) support
pandas>=2.0.0    # Data processing

# Email Processing
mail-parser>=3.15.0  # EML files
beautifulsoup4>=4.12.0  # Web page parsing
requests>=2.31.0

# Audio/Video Processing
moviepy>=1.0.3    # Video processing
SpeechRecognition>=3.10.0  # Speech to text
pydub>=0.25.1     # Audio processing
git+https://github.com/openai/whisper.git    # OpenAI's Whisper for accurate speech recognition

# Enhanced OCR & ML
easyocr>=1.7.1    # Better OCR accuracy
torch>=2.0.0      # Required for EasyOCR
torchvision>=0.15.0
tqdm>=4.66.1      # Progress bars
